Update assets and src info
The hubble image used earlier may have been a reupload by Zolt Levay, but I was feeling unsure about the accreditation, so I replaced it with the NASA original source and much more in-depth source info.
Simple-free-look-camera has been updated. If there's a way to attach it's repo as a dependency, that would probably be better than what I'm doing, but I'll look into that some time later.
I'm hoping to follow something of a standard with my borrowed work, as implemented here. Soon all assets used will follow this format. If the format changes, I aim to change all srcinfo files to accommodate this change.